+++ Core/HLE/sceKernelMutex.cpp
@@ -686,7 +686,7 @@ int sceKernelReferMutexStatus(SceUID id, u32 infoAddr)
 	if (Memory::Read_U32(infoAddr) != 0)
 	{
 		// Refresh and write
-		m->nm.numWaitThreads = m->waitingThreads.size();
+		m->nm.numWaitThreads = (int) m->waitingThreads.size();
 		Memory::WriteStruct(infoAddr, &m->nm);
 	}
 	return 0;
@@ -1133,7 +1133,7 @@ int __KernelReferLwMutexStatus(SceUID uid, u32 infoPtr)
 		// Refresh and write
 		m->nm.currentCount = workarea->lockLevel;
 		m->nm.lockThread = workarea->lockThread == 0 ? -1 : workarea->lockThread;
-		m->nm.numWaitThreads = m->waitingThreads.size();
+		m->nm.numWaitThreads = (int) m->waitingThreads.size();
 		Memory::WriteStruct(infoPtr, &m->nm);
 	}
 	return 0;
